polish up worker.go, downloadheap.go
